+++ pydash/utilities.py
@@ -8,6 +8,7 @@ from __future__ import absolute_import, division
 
 from collections import namedtuple
 from datetime import datetime
+from functools import partial
 import math
 from random import uniform, randint
 import re
@@ -213,8 +214,11 @@ def constant(value):
         True
 
     .. versionadded:: 1.0.0
+
+    .. versionchanged:: TODO
+        Returned function ignores arguments instead of raising exception.
     """
-    return lambda: value
+    return partial(identity, value)
 
 
 def deep_property(path):